requirements.txt
streamlit
pandas
PyPDF2
python-docx
python-dotenv
pymongo
pillow  # Ensures smooth image handling if needed
